fix(router): decomposition route.js
